feat(UIKIT-1122,ui,DashboardLayout): Правки по ПР
